PF-3: Added aboutData to deconstruct ContentSections into page
