+++ src/components/Gallery/Gallery.component.jsx
@@ -1,11 +1,13 @@
 // Import: Dependencies
 import React from "react";
-import styled from "styled-components";
 
-// Import: UI
+// Import: Styled Components
+import { GalleryContainer, GalleryItem } from "./Gallery.elements";
+
+// Import: Components
 import GalleryCard from "../GalleryCard/GalleryCard.component";
 
-// UI: Gallery
+// Components: Gallery
 function Gallery() {
   return (
     <GalleryContainer>
@@ -16,6 +18,7 @@ function Gallery() {
           cardImage="https://images.unsplash.com/photo-1502740479091-635887520276?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1349&q=80"
         />
       </GalleryItem>
+
       <GalleryItem>
         <GalleryCard
           cardText="Acronym Search"
@@ -23,6 +26,7 @@ function Gallery() {
           cardImage="https://images.unsplash.com/photo-1580028468862-6895981d7a67?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80"
         />
       </GalleryItem>
+
       <GalleryItem>
         <GalleryCard
           cardText="Rizji"
@@ -30,6 +34,7 @@ function Gallery() {
           cardImage="https://images.unsplash.com/photo-1550745165-9bc0b252726f?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1950&q=80"
         />
       </GalleryItem>
+
       <GalleryItem>
         <GalleryCard
           cardText="Keyboard Shop"
@@ -37,6 +42,7 @@ function Gallery() {
           cardImage="https://images.unsplash.com/photo-1595225476474-87563907a212?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80"
         />
       </GalleryItem>
+
       <GalleryItem>
         <GalleryCard
           cardText="Portfolio Website"
@@ -50,36 +56,3 @@ function Gallery() {
 
 // Export: Gallery
 export default Gallery;
-
-// Styled: GalleryContainer
-const GalleryContainer = styled.section`
-  align-items: center;
-  background: ${(props) => props.theme.colors.global.backgroundSecondary};
-  display: grid;
-  grid-gap: 1.6rem;
-  grid-template-columns: repeat(2, 1fr);
-  height: auto;
-  margin: auto;
-  padding: 1.6rem;
-  transition: all 150ms linear;
-  width: 70%;
-
-  @media screen and (max-width: 800px) {
-    grid-template-columns: 1fr;
-  }
-
-  @media screen and (max-width: 1024px) {
-    width: 85%;
-  }
-`;
-
-// Styled: GalleryItem
-const GalleryItem = styled.article`
-  height: 40vh;
-
-  @media screen and (min-width: 800px) {
-    &:last-child {
-      grid-column: span 2;
-    }
-  }
-`;
