+++ src/components/index.js
@@ -1,0 +1,20 @@
+// Component: ContentSection
+export { default as ContentSection } from "./ContentSection/ContentSection.component";
+
+// Component: Footer
+export { default as Footer } from "./Footer/Footer.component";
+
+// Component: Gallery
+export { default as Gallery } from "./Gallery/Gallery.component";
+
+// Component: GalleryCard
+export { default as GalleryCard } from "./GalleryCard/GalleryCard.component";
+
+// Component: Header
+export { default as Header } from "./Header/Header.component";
+
+// Component: NavigationMenu
+export { default as NavigationMenu } from "./NavigationMenu/NavigationMenu.component";
+
+// Component: Switch
+export { default as Switch } from "./Switch/Switch.component";
