+++ src/pages/About/aboutData.js
@@ -1,0 +1,39 @@
+// Data: aboutObjOne
+export const aboutObjOne = {
+  isPrimaryBackground: false,
+  heading: "About Me",
+  textOne:
+    "We are a generation of renters. While some people really do prefer the flexibility of private renting, the truth is that for most people it’s just too hard to get on the property ladder. Even before the pandemic, renters were spending 40% of their income on rent (the affordability benchmark is 30%). And, according to the Office of National Statistics, just over a third (36%) of private renters have said they have savings or money invested to buy a property.",
+  textTwo:
+    "For millions of people, we’re in privately rented homes for the long haul. So, with a pandemic and the biggest recession on record dictating our lives, what does it all mean for people living in privately rented properties right now?",
+};
+
+// Data: aboutObjTwo
+export const aboutObjTwo = {
+  isPrimaryBackground: true,
+  heading: "Skills",
+  textOne:
+    "We are a generation of renters. While some people really do prefer the flexibility of private renting, the truth is that for most people it’s just too hard to get on the property ladder. Even before the pandemic, renters were spending 40% of their income on rent (the affordability benchmark is 30%). And, according to the Office of National Statistics, just over a third (36%) of private renters have said they have savings or money invested to buy a property.",
+  textTwo:
+    "For millions of people, we’re in privately rented homes for the long haul. So, with a pandemic and the biggest recession on record dictating our lives, what does it all mean for people living in privately rented properties right now?",
+};
+
+// Data: aboutObjThree
+export const aboutObjThree = {
+  isPrimaryBackground: false,
+  heading: "Work Experience",
+  textOne:
+    "We are a generation of renters. While some people really do prefer the flexibility of private renting, the truth is that for most people it’s just too hard to get on the property ladder. Even before the pandemic, renters were spending 40% of their income on rent (the affordability benchmark is 30%). And, according to the Office of National Statistics, just over a third (36%) of private renters have said they have savings or money invested to buy a property.",
+  textTwo:
+    "For millions of people, we’re in privately rented homes for the long haul. So, with a pandemic and the biggest recession on record dictating our lives, what does it all mean for people living in privately rented properties right now?",
+};
+
+// Data: aboutObjFour
+export const aboutObjFour = {
+  isPrimaryBackground: true,
+  heading: "Education",
+  textOne:
+    "We are a generation of renters. While some people really do prefer the flexibility of private renting, the truth is that for most people it’s just too hard to get on the property ladder. Even before the pandemic, renters were spending 40% of their income on rent (the affordability benchmark is 30%). And, according to the Office of National Statistics, just over a third (36%) of private renters have said they have savings or money invested to buy a property.",
+  textTwo:
+    "For millions of people, we’re in privately rented homes for the long haul. So, with a pandemic and the biggest recession on record dictating our lives, what does it all mean for people living in privately rented properties right now?",
+};
